move validation types to utils
