+++ src/utils/index.ts
@@ -1,3 +1,9 @@
 export { validateFieldType } from "./validateFieldType.ts";
 export { isPerson, validatePeople } from "./validatePeople.ts";
 export { validateUrlTypes } from "./validateUrlTypes.ts";
+export type {
+	People,
+	Person,
+	UrlOrMailTo,
+	UrlType,
+} from "./validation.types.ts";
