+++ src/validators/validateUrlOrMailto.ts
@@ -1,4 +1,4 @@
-import type { UrlOrMailTo } from "./validation.types.ts";
+import type { UrlOrMailTo } from "../utils/index.ts";
 
 import { emailFormat, urlFormat } from "../formats.ts";
